src/rosbag-uploader/pyproject.toml
[tool.poetry]
name = "rosbag-uploader"
version = "0.1.0"
description = ""
authors = ["Georg Teufelberger <tgeorg@student.ethz.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pandas = "^2.2.2"
mysql-connector-python = "^9.0.0"
openpyxl = "^3.1.5"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"


[tool.poetry.group.production]
optional = true

[tool.poetry.group.production.dependencies]
waitress = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
